Move email verification check code to shared
