+++ email-filter-lambda/src/handler.ts
@@ -1,4 +1,5 @@
-import type { SESEvent, SESHandler, SESReceipt } from 'aws-lambda';
+import type { SESEvent, SESHandler } from 'aws-lambda';
+import { findEmailVerificationFailures } from '../../shared/findEmailVerificationFailures';
 import { createLogger } from '../../shared/lambda-logging';
 
 const logger = createLogger({});
@@ -7,7 +8,8 @@ export const main: SESHandler = (event: SESEvent) => {
 	for (const record of event.Records) {
 		const receipt = record.ses.receipt;
 
-		const { hasFailures, failedChecks } = findVerificationFailures(receipt);
+		const { hasFailures, failedChecks } =
+			findEmailVerificationFailures(receipt);
 
 		if (hasFailures) {
 			logger.error({
@@ -22,20 +24,3 @@ export const main: SESHandler = (event: SESEvent) => {
 		console.log('Email verification passed');
 	}
 };
-
-function findVerificationFailures(receipt: SESReceipt) {
-	const verdictChecks = [
-		{ name: 'spamVerdict', status: receipt.spamVerdict.status },
-		{ name: 'virusVerdict', status: receipt.virusVerdict.status },
-		{ name: 'spfVerdict', status: receipt.spfVerdict.status },
-		{ name: 'dkimVerdict', status: receipt.dkimVerdict.status },
-		{ name: 'dmarcVerdict', status: receipt.dmarcVerdict.status },
-	];
-
-	const failedChecks = verdictChecks.filter(({ status }) => status !== 'PASS');
-
-	return {
-		hasFailures: failedChecks.length !== 0,
-		failedChecks,
-	};
-}
